ci: Fix repo settings for self-hosted runners and debugging
The hack of using "secrets" to store per-repo settings was not
working.  The main reason is that pull_request workflows don't have
access to secrets no matter what you do.  So it was impossible to make
this work for settings like "ENABLE_SELF_HOSTED" for PR tests.

This change replaces that old hack with a new one.  Now a repo owner
must create a "GitHub Environment" with the name of the setting they
want to enable.  Currently supported values are "self_hosted", to add
self-hosted runners to the build/test matrix, and "debug", to start an
SSH server for debugging when a workflow fails.