test: Avoid crashing a test executable
LOG(FATAL) aborts the executable, which means gtest can't write any
reports it might be configured to write when tests are complete.  This
will interfere with reporting.

This converts the only use of LOG(FATAL), which was in
ReadTestDataFile(), to LOG(ERROR).  This also updates test cases to
avoid crashing when ReadTestDataFile() returns an empty buffer.