test: Fix bad assertions in formats/mp4 and formats/mp2t tests
PR #1242 introduced some bad assertions that went unnoticed because we
are not yet building those modules in the cmake branch.  This fixes
those so that ASSERT_ macros are not used in non-void functions, and
so that assumptions about a successful parse are always after an
ASSERT_ that the data was loaded properly.