build: Fix rebuild of link-test targets
The link-test targets rebuilt every time you built the project because
that is the behavior of commands attached to add_custom_target.  This
is fixed by combining add_custom_command and add_custom_target.

Fixing that led to a situation where the link-test targets built on
the first build, but didn't properly rebuild when their dependencies
changed.  This is fixed by using set_source_file_properties to treat
test.cc as dirty when the test-install command runs.