build: Separate shared and static library targets
This splits the target so that we always build static libraries, but
optionally also build shared libraries.  This also tweaks the
installation so that the static library is never installed, because it
is not usable without the other internal static library deps.  The
headers and pkgconfig will only be installed if we have a shared
library to use them with.