fix: adapted webvtt unit tests.
The newer, more tolerant header check, allows to ingest webvtt without properly detecting header.
Even in such case, the format autodetection procedure later fails to detect webvtt, so the output for the unit test is still "no streams".
So the only needed change is the failure to parse to a successful parse, but mantaining the empty streams assert.