ci: Read build matrix JSON explicitly
Because we used require() to read build-matrix.json, the file could be replaced with build-matrix.json.js, allowing code injection into our CI pipelines.  This fixes this vulnerability by reading the JSON text with the fs module, then explicitly parsing it, rather than relying on require().

This also changes the location of the file, to match its location in other projects.